+++ packages/entities/entities-plugins/src/components/free-form/AICustomGuardrail/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AICustomGuardrailForm.vue'