+++ packages/analytics/analytics-utilities/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.12.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@12.12.0...@kong-ui-public/analytics-utilities@12.12.1) (2026-03-18)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#3011](https://github.com/Kong/public-ui-components/issues/3011)) ([af2d115](https://github.com/Kong/public-ui-components/commit/af2d115eea61eb0b95036cb4785ee19f4399dad6))
+
+
+
+
+
+# [12.12.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@12.11.0...@kong-ui-public/analytics-utilities@12.12.0) (2026-03-13)
+
+
+### Features
+
+* datasource config store ([#2992](https://github.com/Kong/public-ui-components/issues/2992)) ([2db9601](https://github.com/Kong/public-ui-components/commit/2db960179ece42681cc07fc3e646e6ae04252c15))
+
+
+
+
+
 # [12.11.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-utilities@12.10.4...@kong-ui-public/analytics-utilities@12.11.0) (2026-02-10)